add unit and integration tests

+++ src/main/java/com/yabelova/healthtracker/util/Numbers.java
@@ -1,7 +1,7 @@
 package com.yabelova.healthtracker.util;
 
 /**
- * Парсинг целых чисел из ввода юзера (id из callback-данных, количества доз и т.п.).
+ * Парсинг целых и дробных чисел из ввода юзера (id из callback-данных, количества доз и т.п.).
  * Возвращает {@code null} вместо выброса исключения, если строка не является числом.
  */
 public final class Numbers {
@@ -17,9 +17,9 @@ public final class Numbers {
         }
     }
 
-    public static Long parseLong(String raw) {
+    public static Double parseDouble(String raw) {
         try {
-            return Long.valueOf(raw);
+            return Double.valueOf(raw);
         } catch (NumberFormatException e) {
             return null;
         }
